- Mario animates only when button pressed. So animation properly introduced. - The sprites can now be flipped in the animator, so that it can perform a left and right movement.

+++ Game Engine 2D/FPS.h
@@ -16,7 +16,7 @@ public:
 private:
 	static int frameRate;
 	static int frames;
-	static float speedFactor;
+	static double speedFactor;
 	static int lastTime;
 	static sf::Clock clock;
 };
